+++ scripts/rawimport.groovy
@@ -9,28 +9,19 @@ import com.poc.osm.model.OSMRelation;
 builder.build(osmstream) {
 
 	// défini un flux de sortie, et description de la structure
-	sortie = gdb(path : "c:\\temp\\t.gdb") {
+	sortie = gdb(path : "c:\\temp\\t2.gdb") {
 		featureclass("pts", ESRI_GEOMETRY_POINT,"WGS84") {
-			/* _text("k", size : 40)
-			_integer("mon champ entier")
-			_double("autre champ") */
-			_integer('id')
+			_long('id')
 		}
 		featureclass("lines", ESRI_GEOMETRY_POLYLINE,"WGS84") {
-			/* _text("k", size : 40)
-			_integer("mon champ entier")
-			_double("autre champ") */
-			_integer('id')
+			_long('id')
 		}
 		featureclass("polygon", ESRI_GEOMETRY_POLYGON,"WGS84") {
-			/* _text("k", size : 40)
-			_integer("mon champ entier")
-			_double("autre champ") */
-			_integer('id')
+			_long('id')
 		}
-		table("rels", "WGS84") {
-			_integer('id')
-			_integer('rid')
+		table("rels") {
+			_long('id')
+			_long('rid')
 			_text('role')
 			_text('type')
 		}
@@ -52,26 +43,20 @@ builder.build(osmstream) {
 		transform { OSMRelation e ->
 			e.getFields()?.clear();
 			e.setValue("id",e.getId());
-			
-			f = { entity -> 
-				def listr = new ArrayList<OSMRelatedObject>();
 				
-				entity?.relations.each { OSMRelatedObject r ->
-					
-					OSMAttributedEntity ro = new OSMAttributedEntity(entity.id, entity.fields);
-					ro.setValue("id", entity.id);
-					ro.setValue("rid", r.relatedId);
-					ro.setValue('role', r.relation);
-					ro.setValue('type', r.type);
-					
-					listr << ro;
-					
-				}
+			e?.relations.collect { 
 				
-				listr
+				OSMAttributedEntity ro = new OSMAttributedEntity(e.id, e.fields);
+				ro.setValue("id", e.id);
+				ro.setValue("rid", it.relatedId);
+				ro.setValue('role', it.relation);
+				ro.setValue('type', it.type);
+				
+				ro;
+				
 			}
+				
 			
-			return (f(e) as List);
 		}
 	}