+++ src/main/groovy/com/poc/osm/output/dsl/TBuilder.groovy
@@ -57,6 +57,7 @@ class TBuilder extends FactoryBuilderSupport {
 		registerFactory("_text", new FieldFactory());
 		registerFactory("_integer", new FieldFactory());
 		registerFactory("_double", new FieldFactory());
+		registerFactory("_long", new FieldFactory());
 	}
 
 	/**